+++ src/App.tsx
@@ -25,6 +25,7 @@ import { UserDto } from "./type/dto/user.dto";
 
 import { UserDispatchContext, UserStateContext } from "./context/user-context";
 import requester from "./apis/requester";
+import Timer from "./components/Timer/Timer";
 
 function App() {
   const [user, setUser] = useState<UserDto | undefined>(undefined);
@@ -63,6 +64,7 @@ function App() {
             <Route path="/set-location" element={<EditLocationLevel />} />
             <Route path="/search" element={<SearchPage />} />
             <Route path="/request/:id" element={<RequestPage />} />
+            <Route path="/timer" element={<Timer />} />
             <Route
               path="/*"
               element={<Navigate replace to={token ? "/main" : "/login"} />}
