+++ src/KakaoLogin/OAuth.ts
@@ -1,4 +1,4 @@
 export const CLIENT_ID = "68d70ae34f86a01071c5f8f5d972c593";
-const REDIRECT_URI = "http://localhost:3000/oauth/callback/kakao";
+export const REDIRECT_URI = "http://localhost:3000/oauth/callback/kakao";
 
 export const KAKAO_AUTH_URL = `https://kauth.kakao.com/oauth/authorize?client_id=${CLIENT_ID}&redirect_uri=${REDIRECT_URI}&response_type=code`;
