+++ src/components/Timer/Timer.module.scss
@@ -1,0 +1,66 @@
+.wrapper {
+  position: fixed;
+
+  width: 100%;
+  min-width: 280px;
+  max-width: 450px;
+  height: 100%;
+  min-height: 568px;
+  max-height: 900px;
+  background: white;
+
+  left: 50%;
+  top: 50%;
+  transform: translate(-50%, -50%);
+
+  .back-shadow {
+    visibility: visible;
+    position: absolute;
+    top: 0;
+    width: 100%;
+    height: 100%;
+    background: #777777;
+    opacity: 0.4;
+    z-index: 1;
+    transition-duration: 0.3s;
+  }
+
+  .first-social-login {
+    position: absolute;
+    width: 70%;
+    left: 50%;
+    top: 50%;
+    transform: translate(-50%, -50%);
+
+    background: white;
+    border-radius: 15px;
+
+    z-index: 2;
+
+    display: block;
+    padding: 5%;
+
+    p {
+      text-align: center;
+
+      @media screen and (max-width: 325px) {
+        font-size: 11px;
+      }
+      @media screen and (min-width: 325px) and (max-width: 417px) {
+        font-size: 13px;
+      }
+    }
+    .needed {
+      text-align: left;
+      padding-left: 31px;
+      font-size: 13px;
+
+      @media screen and (max-width: 325px) {
+        font-size: 11px;
+      }
+      @media screen and (min-width: 325px) and (max-width: 417px) {
+        font-size: 12px;
+      }
+    }
+  }
+}
